% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily-wind-plot-from-csv.R
\name{daily_wind_plot}
\alias{daily_wind_plot}
\title{daily_wind_plot}
\usage{
daily_wind_plot(
  csv_file,
  week_number,
  name_plot,
  height,
  width,
  shp_file,
  start_time,
  end_time
)
}
\arguments{
\item{csv_file}{nombre archivo entrada}

\item{week_number}{string que indica la semana a graficar}

\item{name_plot}{string que indica el nombre de la imagen de salida}

\item{height}{altura del plot}

\item{width}{ancho del plot}

\item{shp_file}{nombre del shp utilizado para la imagen final}

\item{start_time}{fecha de inicio del el plot}

\item{end_time}{fecha de termino del plot}
}
\value{
imágenes png de velocidad y magnitud de viento
}
\description{
Función para obtener imágenes png de magnitud y velocidad de viento para datos del ERAS5 (desde archivos csv) Lunes_Viernes
}
\examples{
\dontrun{
csv_file <- "XXXX.csv"
time_step <- 1
week_number <- 12
name_plot <- "wind_plot_semana_12_complete.png"
height <- 11
width <- 8
shp_file <- "/media/evolecolab/PortableSSD/seguimiento_arauco_2022/viento/Chile.shp"
start_time <- "2023-01-23"
end_time <- "2023-01-29"
daily_wind_plot(csv_file = nc_file, time_step = time_step, week_number = week_number, name_plot = name_plot, height = height, width = width, shp_file = shp_file, start_time = start_time, end_time = end_time)
}
}
