% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sentinel2.R
\name{get_cdse_token}
\alias{get_cdse_token}
\title{Obtener un token de acceso de Copernicus Data Space Ecosystem (CDSE)}
\usage{
get_cdse_token(user = Sys.getenv("CDSE_USER"), pass = Sys.getenv("CDSE_PASS"))
}
\arguments{
\item{user}{Cadena. Usuario de CDSE. Por defecto toma la variable de entorno
\code{CDSE_USER}.}

\item{pass}{Cadena. Contraseña de CDSE. Por defecto toma la variable de
entorno \code{CDSE_PASS}.}
}
\value{
Cadena con el \code{access_token} vigente.
}
\description{
Solicita un token OAuth2 al endpoint de identidad de CDSE usando el flujo
\code{password} y el cliente público \code{cdse-public}.
}
\examples{
\dontrun{
token <- get_cdse_token()
}

}
