% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-postproccess.R
\name{get_csv_fix}
\alias{get_csv_fix}
\title{get_csv_fix}
\usage{
get_csv_fix(dir_input, dir_output, var_name, n_cores = 1)
}
\arguments{
\item{dir_input}{directorio en donde se almacenan las imágenes L3}

\item{dir_output}{directorio en donde se almacenaran las imágenes en formato raster}

\item{var_name}{vector de tamaño 1 con el nombre de la variable a analizar ("chlor_a", "sst", "Rrs_645", "pic", "poc", "nflh")}

\item{n_cores}{vector tamaño 1 que indique el numero de núcleos a usar. Por defecto, n_cores = 1}
}
\value{
data.frame csv
}
\description{
Función para generar imágenes raster a partir de imágenes satelitales L3
}
\examples{
\dontrun{
dir_input <- "/home/evolecolab/Escritorio/test_satImg/test_get_L3/chlor_a/"
dir_output <- "/home/evolecolab/Escritorio/test_satImg/test_get_L3/test_raster_fix"
var_name <- "chlor_a"
n_cores <- 4
get_csv_fix(dir_input = dir_input, dir_output = dir_output, var_name = var_name, n_cores = n_cores)
}
}
