% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dataframe.R
\name{get_dataframe}
\alias{get_dataframe}
\title{get_dataframe}
\usage{
get_dataframe(
  dir_input,
  dir_output,
  var_name,
  stat_function = "median",
  custom_time = FALSE,
  begin_day = NULL,
  end_day = NULL,
  n_cores = 1
)
}
\arguments{
\item{dir_input}{directorio en donde se almacenan las imágenes L3 (formato .nc)}

\item{dir_output}{directorio de destino}

\item{var_name}{vector de tamaño 1 con el nombre de la variable a analizar ("chlor_a", "sst", "Rrs_645", "pic", "poc", "nflh")}

\item{stat_function}{función estadística para generar los compuestos diarios ("median", "mean", etc).Por defecto, "median"}

\item{custom_time}{se necesita un intervalo particular de tiempo?. TRUE/FALSE. Si TRUE hay que indicar begin_day y end_day. Si FALSE, se considerara el intervalo de tiempo comprendido por todas las imágenes}

\item{begin_day}{fecha inicio (formato aaaa-mm-dd)}

\item{end_day}{fecha de termino (formato aaaa-mm-dd)}

\item{n_cores}{vector tamaño 1 que indique el numero de núcleos a usar. Por defecto, n_cores = 1}
}
\value{
imágenes raster o archivos parquet
}
\description{
Función para generar imágenes raster o archivos csv a partir de imágenes satelitales L3
}
\examples{
\dontrun{
dir_input <- "/dir/to/input/"
dir_output <- "/dir/to/output/"
stat_function <- "median"
var_name <- "chlor_a"
n_cores <- 4
fecha1 <- "2022-04-18"
fecha2 <- "2022-04-24"
get_dataframe(dir_input = dir_input, dir_output = dir_output, stat_function = stat_function, var_name = var_name, n_cores = n_cores)
}
}
