% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-postproccess.R
\name{get_filled_raster}
\alias{get_filled_raster}
\title{get_filled_raster}
\usage{
get_filled_raster(
  dir_input,
  dir_output,
  shp_mask_file,
  season = "mes",
  n_cores = 1
)
}
\arguments{
\item{dir_input}{directorio en donde se almacenan las imágenes L3}

\item{dir_output}{directorio en donde se almacenaran las imágenes en formato raster}

\item{shp_mask_file}{nombre del archivo usado como mascara (formato .shp)}

\item{season}{temporalidad para la generación de imágenes en formato raster ("semana", "mes"). Actualmente, solo funciona con "mes"}

\item{n_cores}{vector tamaño 1 que indique el numero de núcleos a usar. Por defecto, n_cores = 1}
}
\value{
imágenes raster sin Na
}
\description{
Función basada en el paquete Gapfill para predecir valores Nas en series temporales
}
\examples{
\dontrun{
dir_input <- "/home/evolecol/Escritorio/R_package/test_package/rasters/resultados_raster/"
dir_output <- paste0("/home/evolecol/Escritorio/R_package/test_package", "/", "gapfill")
season <- "mes"
shp_mask_file <- "/home/evolecol/Escritorio/R_package/test_package/Golfo_Arauco_prj2.shp"
n_cores <- 4
get_filled_raster(dir_input = dir_input, dir_output = dir_output, shp_mask_file = shp_mask_file,season = season, n_cores = n_cores)
}
}
