% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-preprocces.R
\name{get_L3}
\alias{get_L3}
\title{get_L3}
\usage{
get_L3(
  dir_input,
  dir_output,
  var_name,
  n_cores = 1,
  res_l2 = "1",
  res_l3 = "1Km",
  north,
  south,
  west,
  east
)
}
\arguments{
\item{dir_input}{directorio en donde se almacenan las imágenes L2}

\item{dir_output}{directorio en donde se almacenaran las imágenes L3 resultantes}

\item{var_name}{nombre de la variable a analizar ("chlor_a", "sst", "Rrs_645", "pic", "poc", "nflh", etc)}

\item{n_cores}{número de núcleos a usar. Por defecto, n_cores = 1 (corrida secuencial)}

\item{res_l2}{resolución para l2bin. Por defecto, res = "1" (H: 0.5km, Q: 250m, HQ: 100m, HH: 50m, 1: 1.1km, 2: 2.3km, 4: 4.6km, 9: 9.2km, 18: 18.5km, 36: 36km, 1D: 1 degree, HD: 0.5 degree, QD: 0.25 degree)}

\item{res_l3}{resolución para l3mapgen. Por defecto, res = "1km" (36km: 1080 x 540, 18km: 2160 x 1080, 9km: 4320 x 2160, 4km: 8640 x 4320, 2km: 17280 x 8640, 1km: 34560 x 17280, hkm: 69120 x 34560, qkm: 138240 x 69120, smi: 4096 x 2048, smi4: 8192 x 4096, land: 8640 x 4320)}

\item{north}{latitud norte para la generación de las imágenes L3}

\item{south}{latitud sur para la generación de las imágenes L3}

\item{west}{latitud oeste para la generación de las imágenes L3}

\item{east}{latitud este para la generación de las imágenes L3}
}
\value{
imágenes L3
}
\description{
Función para obtener imágenes L3 a partir de imágenes L2 descargadas desde https://oceancolor.gsfc.nasa.gov/cgi/browse.pl
}
\examples{
\dontrun{
dir_input <- "/home/evolecolab/Escritorio/test_satImg/test_get_L3"
dir_output <- "/home/evolecolab/Escritorio/test_satImg/test_get_L3"
var_name <- "chlor_a"
n_cores <- 6
res_l2 <- "1"
res_l3 <- "1km"
north <- -36.77544
south <- -37.29073
west <- -73.67165
east <- -73.11573
get_L3(dir_input = dir_input, dir_output = dir_output, var_name = var_name, n_cores = n_cores, res_l2=res_l2, res_l3 = res_l3, north = north, sout = south, west = west, east = east)
}
}
