% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-raster-csv.R
\name{get_raster_csv}
\alias{get_raster_csv}
\title{get_raster_csv}
\usage{
get_raster_csv(
  dir_input,
  dir_output,
  season = "month",
  raster_function = "median",
  result_type,
  var_name,
  n_cores = 1
)
}
\arguments{
\item{dir_input}{directorio en donde se almacenan las imágenes L3 (formato .nc)}

\item{dir_output}{directorio de destino}

\item{season}{temporalidad para la generación de imágenes en formato raster. Opciones "year", "month", "week"}

\item{raster_function}{función estadística para generar los compuestos ("median" o "mean").  Por defecto, median}

\item{result_type}{tipo de salida requerida. Si result_type == "raster", se generaran imágenes raster según la estacionalidad y función para los compuestos definidos. Si result_type =="csv", se generar archivos año/mes en fomato csv}

\item{var_name}{vector de tamaño 1 con el nombre de la variable a analizar ("chlor_a", "sst", "Rrs_645", "pic", "poc", "nflh")}

\item{n_cores}{vector tamaño 1 que indique el numero de núcleos a usar. Por defecto, n_cores = 1}
}
\value{
imágenes raster o archivos csv
}
\description{
Función para generar imágenes raster o archivos csv a partir de imágenes satelitales L3
}
