% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-postproccess.R
\name{get_raster_ct}
\alias{get_raster_ct}
\title{get_raster_ct}
\usage{
get_raster_ct(
  dir_input,
  dir_output,
  date_1,
  date_2,
  name_time,
  var_name,
  raster_function = "median"
)
}
\arguments{
\item{dir_input}{directorio en donde se almacenas las imágenes L3}

\item{dir_output}{directorio en donde se almacenaras las imágenes en formato raster}

\item{date_1}{fecha inicio del tiempo personalizado (formato: AAAA-MM-DD)}

\item{date_2}{fecha termino del tiempo personalizado (formato: AAAA-MM-DD)}

\item{name_time}{vector tamaño 1 que indica el nombre del intervalo temporal ("eg. "semana 1")}

\item{var_name}{vector de tamaño 1 con el nombre de la variable a analizar ("chlor_a", "sst", "Rrs_645", "pic", "poc", "nflh")}

\item{raster_function}{función estadística para generar las imágenes raster ("median" o "mean")}
}
\value{
imágenes raster
}
\description{
Función para generar imágenes raster a partir de imágenes L3 para un intervalo de tiempo determinado (desde intervalos diarios en adelante)
}
\examples{
\dontrun{
dir_input <- "/home/evolecol/Escritorio/R_package/test_package/chlor_A/"
dir_output <- paste0( "/home/evolecol/Escritorio/R_package/test_package/test_ct")
date_1 <- "2010-01-01"
date_2 <- "2010-01-29"
name_time <- "enero_2010"
var_name <- "chlor_a"
raster_function <- "median"
get_raster_ct(dir_input = dir_input, dir_output = dir_output, date_1 = date_1, date_2 = date_2, name_time = "caca_2010",var_name = "chlor_a",raster_function = raster_function)
}
}
