% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-postproccess.R
\name{get_raster_fix}
\alias{get_raster_fix}
\title{get_raster_fix}
\usage{
get_raster_fix(
  dir_input,
  dir_output,
  season = "mes",
  raster_function = "median",
  var_name,
  n_cores = 1
)
}
\arguments{
\item{dir_input}{directorio en donde se almacenan las imágenes L3}

\item{dir_output}{directorio en donde se almacenaran las imágenes en formato raster}

\item{season}{temporalidad para la generación de imágenes en formato raster ("semana", "mes", año). Se obtiene 1 raster por temporalidad (e.g. 1 raster por cada semana, o un raster por cada mes). Por defecto "mes"}

\item{raster_function}{función estadística para generar las imágenes raster ("median" o "mean").  Por defecto, median}

\item{var_name}{vector de tamaño 1 con el nombre de la variable a analizar ("chlor_a", "sst", "Rrs_645", "pic", "poc", "nflh")}

\item{n_cores}{vector tamaño 1 que indique el numero de núcleos a usar. Por defecto, n_cores = 1}
}
\value{
Imágenes raster
}
\description{
Función para generar imágenes raster a partir de imágenes satelitales L3 (formato aaajulianday)
}
\examples{
\dontrun{
dir_input <- "/home/evolecolab/Escritorio/test_satImg/test_get_L3/chlor_a/"
dir_output <- "/home/evolecolab/Escritorio/test_satImg/test_get_L3/test_raster_fix"
season <- "mes"
raster_function <- "median"
var_name <- "chlor_a"
n_cores <- 4
get_raster_fix(dir_input = dir_input, dir_output = dir_output, season = season,raster_function = raster_function,var_name = var_name,n_cores = n_cores)
}
}
