% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tile_grid}
\alias{get_tile_grid}
\title{Obtener la grilla de tiles}
\usage{
get_tile_grid(refresh = FALSE)
}
\arguments{
\item{refresh}{Lógico. Si \code{TRUE}, fuerza la recarga desde disco.}
}
\value{
Un objeto \code{sf}.
}
\description{
Carga la grilla de tiles desde los archivos del paquete.
El objeto se mantiene en memoria una vez cargado para optimizar el rendimiento.
}
