% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-wind-data-full.R
\name{get_wind_data}
\alias{get_wind_data}
\title{get_wind_data}
\usage{
get_wind_data(
  lat_min,
  lat_max,
  long_min,
  long_max,
  start_time,
  end_time,
  dir_output = getwd(),
  product_name,
  prefix_outfile
)
}
\arguments{
\item{lat_min}{latitud mínima para la zona de descarga}

\item{lat_max}{latitud máxima para la zona de descarga}

\item{long_min}{latitud mínima para la zona de descarga}

\item{long_max}{longitud máxima para la zona de descarga}

\item{start_time}{tiempo inicial para la descarga}

\item{end_time}{tiempo dinal para la descarga}

\item{dir_output}{directorio de descarga de los datos. Por defecto(getwd())}

\item{product_name}{tipo de producto. Por ahora 2: "pressure-levels" o "single-levels"}

\item{prefix_outfile}{archivo a descargar}
}
\value{
dataframe en .csv
}
\description{
Función para descarcar datos de viento desde eras5 (u y v) para 1000 hpa (formato .grib), por medio del paquete ecmwfr y transformarlos a un dataframe (csv) diario. Es necesario tener una cuenta del CDS.
}
\examples{
\dontrun{
long_max <- -60
long_min <- -80
lat_max <- -18
lat_min <- -56
start_time <- "2023-02-20"
end_time <- "2023-02-24"
dir_output <- dir_output
prefix_outfile <- prefix_outfile
uid <- "XXXXX"
cds_api_key <- "XXXXXX"
ecmwfr::wf_set_key(
  user = uid,
  key = cds_api_key,
  service = "cds"
)
get_wind_data(
  lat_min = lat_min, lat_max = lat_max, long_min = long_max, star_time = star_time,
  end_time = end_time, dir_output = dir_output, prefix_outfile = prefix_outfile
)
}
}
