% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-wind-plot-from-nc.R
\name{get_wind_plot}
\alias{get_wind_plot}
\title{get_wind_plot}
\usage{
get_wind_plot(
  nc_file,
  time_step,
  week_number,
  name_plot,
  height,
  width,
  shp_file,
  fecha_1,
  fecha_2
)
}
\arguments{
\item{nc_file}{nombre archivo entrada}

\item{time_step}{parametro que indica la temporalidad de los datos de ERAS 5 (defecto = 1; cada una hora)}

\item{week_number}{string que indica la semana a graficar}

\item{name_plot}{string que indica el nombre de la imagen de salida}

\item{height}{altura del plot}

\item{width}{ancho del plot}

\item{shp_file}{nombre del shp utilizado para la imagen final}

\item{fecha_1}{fecha de inicio del el plot}

\item{fecha_2}{fecha de termino del plot}
}
\value{
imágenes png de velocidad y magnitud de viento
}
\description{
Función para obtener imágenes png de magnitud y velocidad de viento para datos del ERAS5 (formato .nc)
}
\examples{
\dontrun{
nc_file <- "wind_20023_ene_w12c.nc"
time_step <- 1
week_number <- 12
name_plot <- "wind_plot_semana_12_complete.png" 
height = 11
width = 8
shp_file <- "/media/evolecolab/PortableSSD/seguimiento_arauco_2022/viento/Chile.shp"
fecha_1 <- "2023-01-23"
fecha_2 <- "2023-01-29"
get_wind_plot(nc_file = nc_file,time_step = time_step, week_number = week_number,name_plot = name_plot, height = height, width = width,shp_file = shp_file, fecha_1 = fecha_1, fecha_2 = fecha_2)
}
}
