% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getL3-1P.R
\name{getL3_1P}
\alias{getL3_1P}
\title{getL3_1P}
\usage{
getL3_1P(
  dir_ocssw,
  dir_input,
  dir_output,
  var_name,
  n_cores = 1,
  res_l2 = "1",
  res_l3 = "1Km",
  north,
  south,
  west,
  east,
  need_extract_and_format = TRUE,
  sort_files = FALSE,
  fudge,
  area_weighting,
  begin_day,
  end_day,
  period_name
)
}
\arguments{
\item{dir_ocssw}{directory en donde estan los binarios ocssw (seadas)}

\item{dir_input}{directory en donde se almacenan las imágenes L2}

\item{dir_output}{directory en donde se almacenaran las imágenes L3 resultantes}

\item{var_name}{nombre de la variable a analizar ("chlor_a", "sst", "Rrs_645", "pic", "poc", "nflh", etc)}

\item{n_cores}{número de núcleos a usar. Por defecto, n_cores = 1 (corrida secuencial). La parelelización es respecto de la cantidad de sub_folder procesados simultaneamente}

\item{res_l2}{resolución para l2bin. Por defecto, res = "1" (H: 0.5km, Q: 250m, HQ: 100m, HH: 50m, 1: 1.1km, 2: 2.3km, 4: 4.6km, 9: 9.2km, 18: 18.5km, 36: 36km, 1D: 1 degree, HD: 0.5 degree, QD: 0.25 degree)}

\item{res_l3}{resolución para l3mapgen. Por defecto, res = "1km" (36km: 1080 x 540, 18km: 2160 x 1080, 9km: 4320 x 2160, 4km: 8640 x 4320, 2km: 17280 x 8640, 1km: 34560 x 17280, hkm: 69120 x 34560, qkm: 138240 x 69120, smi: 4096 x 2048, smi4: 8192 x 4096, land: 8640 x 4320)}

\item{north}{latitud norte para la generación de las imágenes L3}

\item{south}{latitud sur para la generación de las imágenes L3}

\item{west}{latitud oeste para la generación de las imágenes L3}

\item{east}{latitud este para la generación de las imágenes L3}

\item{sort_files}{crearr sistema de archivos año/mes? (TRUE/FALSE).Por defecto, FALSE}

\item{fudge}{??fusionador}

\item{area_weighting}{area_weighting}

\item{begin_day}{fecha inicio de temporalidad(ymd)}

\item{end_day}{fecha temporalidad de temporalidad (ymd)}

\item{period_name}{cadena texto temporalidad a usar (e.g "semana_01")}

\item{need_extract}{mantener sistema de archivos año/mes? (TRUE/FALSE).Por defecto, FALSE}

\item{flaguse}{cadena de texto con los nosmbres de los flag a usar}
}
\value{
imágenes L3
}
\description{
Función para obtener imágenes L3 a partir de imágenes L2 descargadas desde https://oceancolor.gsfc.nasa.gov/cgi/browse.pl
}
\examples{
\dontrun{
dir_input <- "/home/evolecol/Desktop/R_package/test_functions/test_new_L3/data_raw"
dir_output <- "/home/evolecol/Desktop/R_package/test_functions/test_new_L3"
dir_ocssw <- "/home/evolecol/SeaDAS/ocssw"
var_name <- "sst"
n_cores <- 6
res_l2 <- "1"
res_l3 <- "1km"
north <- -35
south <- -40
west <- -75
east <- -70
flaguse <- "LAND,HISOLZEN"
need_extract_and_format <- TRUE
sort_files <- FALSE
period_name <- "w_08"
area_weighting <- 0
fudge <- 2
begin_day <- "2022-12-26"
end_day <- "2022-12-28"
getL3_1P(dir_ocssw = dir_ocssw, dir_input = dir_input, dir_output = dir_output, var_name = var_name, n_cores = n_cores, res_l2 = res_l2, res_l3 = res_l3, north = north, south = south, west = west, east = east, need_extract_and_format = need_extract_and_format, sort_files = sort_files, begin_day = begin_day, end_day = end_day, period_name = period_name)
}
}
