% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2-to-dataframe.R
\name{l2_to_dataframe}
\alias{l2_to_dataframe}
\title{l2_to_dataframe}
\usage{
l2_to_dataframe(
  dir_ocssw,
  dir_input,
  dir_output,
  format_output = "parquet",
  sensor,
  var_name,
  season,
  n_cores = 1,
  res_l2 = "1",
  res_l3 = "1Km",
  north,
  south,
  west,
  east,
  fudge,
  area_weighting = 0,
  data_compress = TRUE
)
}
\arguments{
\item{dir_ocssw}{Directorio OCSSW (SeaDAS)}

\item{dir_input}{Directorio con archivos L2 (o .tar si data_compress = TRUE)}

\item{dir_output}{Directorio de salida para L3 y archivos finales}

\item{format_output}{Formato de salida para las tablas ('parquet' por defecto)}

\item{sensor}{Sensor MODIS/Sentinel ('aqua', 'terra', etc.)}

\item{var_name}{Nombre del producto (e.g. 'chlor_a')}

\item{season}{Periodo de agregación ('year', 'month', 'day')}

\item{n_cores}{Núm. de núcleos para procesamiento paralelo}

\item{res_l2}{Resolución de binning (default '1')}

\item{res_l3}{Resolución de mapeo ('1Km')}

\item{north, }{south, west, east Límites espaciales}

\item{fudge}{Valor fudge para mapgen}

\item{area_weighting}{Si se aplica ponderación por área}

\item{data_compress}{Si TRUE, espera archivos .tar; si FALSE, archivos .nc directos}
}
\description{
Procesa archivos L2 (SeaDAS) y los transforma a L3 y luego a tabla (e.g., parquet). Los archivos L2 y temporales permanecen en \code{dir_input}; solo los resultados finales en parquet o csv se almacenan en \code{dir_output}.
}
