% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l3-to-dataframe.R
\name{l3_to_dataframe}
\alias{l3_to_dataframe}
\title{l3_to_dataframe}
\usage{
l3_to_dataframe(
  dir_input,
  dir_output,
  season,
  var_name,
  star_date = NULL,
  end_date = NULL,
  north = NULL,
  south = NULL,
  west = NULL,
  east = NULL,
  stat_function = "mean",
  format_output,
  n_cores = 1
)
}
\arguments{
\item{dir_input}{directorio en donde se almacenan las imágenes L3}

\item{dir_output}{directorio en donde se almacenaran los archivos resultantes}

\item{season}{temporalidad para la generación de las tablas de datos ("day",  "month", "year" por ahora). Depende la la temporalidad de las imágenes L3}

\item{var_name}{nombre de la variable a analizar ("chlor_a", "sst", "Rrs_645", "pic", "poc", "nflh", etc)}

\item{north}{latitud norte. Si null, se toma la latitud norte de los datos de entrada}

\item{south}{latitud sur. Si null, se toma la latitud sur de los datos de entrada}

\item{west}{latitud oeste. Si null, se toma la longitud oeste de los datos de entrada}

\item{east}{latitud este. Si null, se toma la longitud este de los datos de entrada}

\item{stat_function}{string con el nombre de la función para calcular el valor de una variables a través de la temporalidad (season).}

\item{format_output}{formato de los archivos resultantes. Toma valores "parquet" o "csv. Por defecto parquet.}

\item{n_cores}{número de núcleos a usar. Por defecto, n_cores = 1 (corrida secuencial). La parelelización es respecto de la cantidad de dataframe procesados simultáneamente}
}
\value{
tabla de datos en csv o parquet
}
\description{
Función para obtener tablas de datos en parquet o csv a partir de imágenes L3 ya procesadas por l2_to_dataframe. Útil cuando se quiere obtener solo una parte (temporal y/o espacial) de los datos
}
\examples{
\dontrun{
dir_input <- "vars_oce/sst/L3"
dir_output <- getwd()
season <- "month"
start_date <- NULL
end_date <- NULL
var_name <- "sst"
north <- -36
south <- -38
west <- -73
east <- -71
stat_function <- "median"
format_output <- "parquet"
l3_to_dataframe(dir_input = dir_input, dir_output = dir_output, season = season, format_ouput = format_ouput, var_name = var_name,  north = north, south = south, west = west, east = east, stat_function = stat_function)
}
}
