% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LonLabel}
\alias{LonLabel}
\alias{LatLabel}
\title{LonLabel}
\usage{
LonLabel(lon, east = "°E", west = "°W", zero = "°")

LatLabel(lat, north = "°N", south = "°S", zero = "°")
}
\description{
LonLabel

LatLabel
}
\keyword{internal}
