% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-facet-oce.R
\name{plot_facet_oce}
\alias{plot_facet_oce}
\title{plot_facet_oce}
\usage{
plot_facet_oce(
  list_nc,
  var,
  name_plot,
  height,
  width,
  shp_file,
  start_time,
  end_time
)
}
\arguments{
\item{var}{string que indica el nombre de la variable a graficar}

\item{name_plot}{string que indica el nombre de la imagen de salida}

\item{height}{altura del plot}

\item{width}{ancho del plot}

\item{shp_file}{nombre del shp utilizado para la imagen final}

\item{start_time}{fecha de inicio del el plot}

\item{end_time}{fecha de termino del plot}

\item{lists_nc}{lista de archivos nc de las variables}
}
\value{
imágenes png parámetros oceanográficos
}
\description{
Función para obtener imágenes del informe mensual para cada parametro oceanográfico (chlor_a, sst y rrs_645)
}
