% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-week-oce.R
\name{plot_week_oce}
\alias{plot_week_oce}
\title{plot_week_oce}
\usage{
plot_week_oce(
  lists_nc,
  week_number,
  name_plot,
  height,
  width,
  shp_file,
  start_time,
  end_time
)
}
\arguments{
\item{lists_nc}{lista de archivos nc de las variables}

\item{week_number}{string que indica la semana a graficar}

\item{name_plot}{string que indica el nombre de la imagen de salida}

\item{height}{altura del plot}

\item{width}{ancho del plot}

\item{shp_file}{nombre del shp utilizado para la imagen final}

\item{start_time}{fecha de inicio del el plot}

\item{end_time}{fecha de termino del plot}
}
\value{
imágenes png parámetros oceanográficos
}
\description{
Función para obtener imágenes semanales de parametros oceanográficos (chlor_a, sst y rrs_645)
}
\examples{
\dontrun{
shp_file <- "/media/evolecolab/PortableSSD/seguimiento_arauco_2022/oce_pars/Golfo_Arauco_prj2.shp"
height <-  5
width <-  14
lists_nc <- dir_ls(path = "/media/evolecolab/PortableSSD/seguimiento_arauco_2022/oce_pars/vars/input_plot/semana_20",regexp = ".nc$")
week_number <- 20
start_time <- "2023-03-20"
end_time <- "2023-03-26"
name_plot <- "ocepars_plot_w20.png" 
plot_week_oce(lists_nc = lists_nc, week_number = week_number,name_plot = name_plot,height = height, width = width,shp_file = shp_file,start_time  = start_time, end_time = end_time)
}
}
