% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-wind.R
\name{plot_wind}
\alias{plot_wind}
\title{plot_wind}
\usage{
plot_wind(
  data,
  start_date,
  end_date,
  cell_size = 150,
  shp_sf,
  n_col,
  xlim,
  ylim,
  output_name,
  width,
  height
)
}
\arguments{
\item{data}{Dataframe con columnas \code{lon}, \code{lat}, \code{date}, \code{u}, \code{v}, \code{speed_mean}.}

\item{start_date}{Fecha de inicio del periodo (formato "YYYY-MM-DD").}

\item{end_date}{Fecha de término del periodo (formato "YYYY-MM-DD").}

\item{cell_size}{Tamaño de las celdas para agregación espacial (por defecto 150).}

\item{shp_sf}{Objeto \code{sf} con los polígonos del shapefile de fondo (ya transformado a CRS 4326).}

\item{n_col}{Número de columnas para la grilla de facetas.}

\item{xlim}{Vector con límites longitudinales (e.g., c(-76, -71.5)).}

\item{ylim}{Vector con límites latitudinales (e.g., c(-38, -36)).}

\item{output_name}{Nombre del archivo de salida (debe terminar en .png).}

\item{width}{Ancho del archivo en pulgadas.}

\item{height}{Alto del archivo en pulgadas.}
}
\value{
Exporta un gráfico \code{.png} que muestra la velocidad y dirección promedio del viento por semana.
}
\description{
Genera un gráfico semanal de velocidad y dirección promedio del viento sobre una región específica.
}
\examples{
\dontrun{
data <- arrow::read_parquet("data_viento_2025.parquet")
shp_sf <- poligon.shp
plot_wind(
  data = data,
  start_date = "2025-01-01",
  end_date = "2025-01-31",
  cell_size = 150,
  shp_sf = shp_sf,
  n_col = 2,
  xlim = c(-76, -71.5),
  ylim = c(-38, -36),
  output_name = "plot.png",
  width = 7,
  height = 10
)
}
}
