% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1-to-l2.R
\name{process_l1a}
\alias{process_l1a}
\title{Procesar un granulo del l1a a L2}
\usage{
process_l1a(
  l1a_file,
  dir_ocssw,
  out_dir,
  north = -36.3,
  south = -36.5,
  west = -73.1,
  east = -72.25,
  resolution = 250,
  limpiar = TRUE
)
}
\arguments{
\item{l1a_file}{L1A (.L1A_LAC).}

\item{dir_ocssw}{Directorio raiz de OCSSW.}

\item{out_dir}{Directorio de salida (donde queda el L2).}

\item{north, south, west, east}{Caja lat/lon.}

\item{resolution}{Resolucion en metros (250).}

\item{limpiar}{Si TRUE, borra la carpeta de trabajo al terminar.}
}
\value{
Ruta del L2 generado.
}
\description{
Procesar un granulo del l1a a L2
}
