% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_y_latitude}
\alias{scale_y_latitude}
\title{scale_x_latitude}
\usage{
scale_y_latitude(
  name = "",
  ticks = 30,
  breaks = seq(-90, 90, by = ticks),
  expand = c(0, 0),
  labels = LatLabel,
  ...
)
}
\description{
scale_x_latitude
}
\keyword{internal}
