% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_x_longitude}
\alias{scale_x_longitude}
\title{scale_x_longitude}
\usage{
scale_x_longitude(
  name = "",
  ticks = 30,
  breaks = seq(-180, 360, by = ticks),
  expand = c(0, 0),
  labels = LonLabel,
  trans = "identity",
  ...
)
}
\description{
scale_x_longitude
}
\keyword{internal}
