% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seadas_l2bin}
\alias{seadas_l2bin}
\title{seadas_l2bin}
\usage{
seadas_l2bin(infile, ofile)
}
\arguments{
\item{infile}{an input file}

\item{ofile}{an input file}
}
\description{
seadas_l2bin
}
\keyword{internal}
