% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seadas_l3bin}
\alias{seadas_l3bin}
\title{seadas_l3bin}
\usage{
seadas_l3bin(infile, ofile)
}
\arguments{
\item{infile}{an input file}

\item{ofile}{an input file}
}
\description{
seadas_l3bin
}
\keyword{internal}
