% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seadas_l3mapgen}
\alias{seadas_l3mapgen}
\title{seadas_l3mapgen}
\usage{
seadas_l3mapgen(infile, ofile)
}
\arguments{
\item{infile}{an input file}

\item{ofile}{an input file}
}
\description{
seadas_l3mapgen
}
\keyword{internal}
