% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toc}
\alias{toc}
\alias{nc_to_table}
\alias{write_table}
\title{toc}
\usage{
toc()

nc_to_table(file)

write_table(df, file)
}
\arguments{
\item{file}{as input}

\item{df}{as input}

\item{msg}{as input}
}
\description{
toc

write_table

nc_to_table
}
\keyword{internal}
