% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly-wind-plot-from-csv.R
\name{weekly_wind_plot}
\alias{weekly_wind_plot}
\title{weekly_wind_plot}
\usage{
weekly_wind_plot(
  list_csv,
  name_plot,
  shp_file,
  start_time,
  end_time,
  height,
  width,
  ncol
)
}
\arguments{
\item{list_csv}{lista con los archivos csv para el plot}

\item{name_plot}{string que indica el nombre de la imagen de salida}

\item{shp_file}{nombre del shp utilizado para la imagen final}

\item{start_time}{fecha de inicio del el plot}

\item{end_time}{fecha de termino del plot}

\item{height}{altura del plot}

\item{width}{ancho del plot}

\item{ncol}{numero de columnas para el plot}
}
\value{
imágenes png de velocidad y magnitud de viento
}
\description{
Función para obtener imágenes png de magnitud y velocidad de viento para datos del ERAS5 (desde archivos csv) para n semanas
}
\examples{
\dontrun{
list_csv <- "list/files/csv"
name_plot <- "wind_plot_semana_12_complete.png"
height <- 11
width <- 8
shp_file <- "/media/evolecolab/PortableSSD/seguimiento_arauco_2022/viento/Chile.shp"
start_time <- "2023-01-23"
end_time <- "2023-01-29"
ncol <- 2
weekly_wind_plot(list_csv = list_csv, name_plot = name_plot, height = height, width = width, shp_file = shp_file, start_time = start_time, end_time = end_time, ncol = ncol)
}
}
