% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-clim.R
\name{get_clim}
\alias{get_clim}
\title{get_clim}
\usage{
get_clim(
  dir_input,
  dir_output,
  season,
  stat_function,
  var_name,
  shp_file,
  n_col,
  n_row,
  name_output,
  n_cores = 1,
  res = 300,
  height = 8,
  width = 6,
  xlim,
  ylim
)
}
\arguments{
\item{dir_input}{directorio en donde se almacenan las imágenes raster (formato .tif)}

\item{dir_output}{directorio en donde se almacenará la imagen png}

\item{season}{temporalidad para la generación de imágenes en formato raster("mes", o "year")}

\item{stat_function}{función estadística para generar las imágenes raster ("median" o "mean")}

\item{var_name}{nombre de la variable a analizar ("chlor_a", "sst", "Rrs_645", "pic", "poc", "nflh", etc)}

\item{shp_file}{nombre archivo shp para el gráfico (si no esta en dir_input poner nombre con ruta completa)}

\item{n_col}{numero de columnas para el gráfico}

\item{n_row}{numero de filas para el gráfico}

\item{name_output}{nombre para la salidas}

\item{res}{resolución para la imágen png de la climatología}

\item{height}{altura para la imágen png de la climatología}

\item{width}{amplitud para la imágen png de la climatología}

\item{xlim}{vector numérico tamaño 2 para los limites de x en el plot}

\item{ylim}{vector numérico tamaño 2 para los limites de y en el plot}
}
\value{
imágenes png y tif de climatologías y Rdata
}
\description{
Función para generar climatología y exportarla (formato png) para una variable determinada
}
\examples{
\dontrun{
dir_input <- "/dir/to/input/"
dir_output <- "/dir/to/ouput/"
season <- "month"
stat_function <- "median"
var_name <- "chlor_a"
n_col <- 3
n_row <- 4
name_output <- "climatologia_chlor_a.png"
shp_file <- "/home/evolecolab/Escritorio/test_satImg/raster_mensuales/resultados_raster/Golfo_Arauco_prj2.shp"
res <- 300
  height <- 7
width <- 9
get_clim(dir_input = dir_input, dir_output = dir_output, season = season, stat_function = stat_function, var_name = var_name, shp_file = shp_file, n_col = n_col, n_row = n_row, name_output = name_output, res = res,   height =   height, width = width)
}
}
