% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dataframe-new.R
\name{get_dataframe}
\alias{get_dataframe}
\title{get_dataframe}
\usage{
get_dataframe(
  dir_input,
  dir_output,
  type,
  var_name,
  stat_function = "median",
  season,
  n_cores = 4
)
}
\arguments{
\item{dir_input}{directorio en donde se almacenan las imágenes L3 (formato .nc)}

\item{dir_output}{directorio de destino}

\item{type}{tipo de dataframe(to_plot o full_df)}

\item{var_name}{vector de tamaño 1 con el nombre de la variable a analizar ("chlor_a", "sst", "Rrs_645", "pic", "poc", "nflh")}

\item{season}{vector de tamaño 1 con la temporalidad deseada para el dataframe("year" or "month" por ahora)}

\item{n_cores}{vector tamaño 1 que indique el numero de núcleos a usar. Por defecto, n_cores = 1}
}
\value{
archivos parquet
}
\description{
Función para generar dataframes en formato .parquet a partir de imágenes satelitales L3
}
\examples{
\dontrun{
dir_input <- "/media/evolecolab/PortableSSD/Proyecto_Packard/chlor_a/join_l3"
dir_output <- "/home/evolecolab/Escritorio/proyecto_packard/chlor_a/to_plot"
type <- "to_plot"
var_name <- "chlor_a"
stat_function <- "median"
season <- "month"
n_cores <- 8
get_dataframe(dir_input = dir_input, dir_output = dir_output, type = "to_plot",var_name = var_name, stat_function = stat_function, season = season,n_cores = n_cores)
}
}
