% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-raster-csv.R
\name{get_raster_csv}
\alias{get_raster_csv}
\title{get_raster_csv}
\usage{
get_raster_csv(
  dir_input,
  dir_output,
  season = "month",
  result_type,
  var_name,
  n_cores = 1,
  stat_function = "median",
  begin_day = NULL,
  end_day = NULL
)
}
\arguments{
\item{dir_input}{directorio en donde se almacenan las imágenes L3 (formato .nc)}

\item{dir_output}{directorio de destino}

\item{season}{temporalidad para la generación de imágenes en formato raster. Opciones "year", "month", "week"}

\item{result_type}{tipo de salida requerida. Si result_type == "raster", se generaran imágenes raster según la estacionalidad y función para los compuestos definidos. Si result_type == "data_frame", se generar archivos año/mes en fomato .parquet}

\item{var_name}{vector de tamaño 1 con el nombre de la variable a analizar ("chlor_a", "sst", "Rrs_645", "pic", "poc", "nflh")}

\item{n_cores}{vector tamaño 1 que indique el numero de núcleos a usar. Por defecto, n_cores = 1}

\item{stat_function}{función estadística para generar los compuestos ("median", "mean", etc).Por defecto, "median"}

\item{begin_day}{fecha inicio (formato aaaa-mm-dd)}

\item{end_day}{fecha de termino (formato aaaa-mm-dd)}

\item{custom_time}{se necesita un intervalo particular de tiempo?. TRUE/FALSE. Si TRUE hay que indicar begin_day y end_day. Si FALSE, se considerara el intervalo de tiempo comprendido por todas las imágenes}
}
\value{
imágenes raster o archivos parquet
}
\description{
Función para generar imágenes raster o archivos csv a partir de imágenes satelitales L3
}
\examples{
\dontrun{
dir_input <- "/dir/to/input/"
dir_output <- "/dir/to/output/"
season <- "month"
stat_function <- "median"
var_name <- "chlor_a"
result_type <- "raster"
n_cores <- 4
fecha1 <- "2022-04-18"
fecha2 <- "2022-04-24"
get_raster_csv_ct(dir_input = dir_input, dir_output = dir_output, season = season, stat_function = stat_function, var_name = var_name, n_cores = n_cores, result_type = "raster")
}
}
