% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-data.R
\name{get_tile_grid}
\alias{get_tile_grid}
\title{Cargar (y cachear) la grilla de tiles Sentinel-2}
\usage{
get_tile_grid(kml)
}
\arguments{
\item{kml}{Cadena. Ruta al archivo KML/FGB con la grilla de tiles.}

\item{refresh}{Lógico. Si \code{TRUE}, fuerza la relectura desde disco.}
}
\value{
Objeto \code{sf} con columnas \code{Name} y \code{geometry}.
}
\description{
Cargar (y cachear) la grilla de tiles Sentinel-2
}
