% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-wind-data.R
\name{get_wind_data}
\alias{get_wind_data}
\title{get_wind_data}
\usage{
get_wind_data(
  lat_min,
  lat_max,
  long_min,
  long_max,
  start_time,
  end_time,
  dir_output,
  name_outfile
)
}
\arguments{
\item{lat_min}{latitud mínima para la zona de descarga}

\item{lat_max}{latitud máxima para la zona de descarga}

\item{long_min}{latitud mínima para la zona de descarga}

\item{long_max}{longitud máxima para la zona de descarga}

\item{start_time}{tiempo inicial para la descarga}

\item{end_time}{tiempo dinal para la descarga}

\item{dir_output}{directorio de descarga de los datos. Por defecto(getwd())}

\item{name_outfile}{nombre archivo a descargar}
}
\value{
archivos .grib
}
\description{
Función para descarcar datos de viento desde eras5 (u y v) para 1000 hpa (formato .grib), por medio del paquete ecmwfr. Es necesario tener una cuenta del CDS.
}
\examples{
\dontrun{
long_max <- -60
long_min <- -80
lat_max <- -18
lat_min <- -56
start_time <- "2023-02-20"
end_time <- "2023-02-24"
out_path <- getwd()
name_outfile <- "data_feb_2023_w15c_w16p.grib"
uid <- "XXXXX"
cds_api_key <- "XXXXXX"
ecmwfr::wf_set_key(
  user = uid,
  key = cds_api_key,
  service = "cds"
)
get_wind_data(
  lat_min = lat_min, lat_max = lat_max, long_min = long_max, star_time = star_time,
  end_time = end_time, dir_output = dir_output, name_outfile = name_outfile
)
}
}
