% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-clim.R
\name{plot_clim}
\alias{plot_clim}
\title{plot_clim}
\usage{
plot_clim(
  dir_input = NULL,
  season,
  stat_function,
  var_name,
  shp_file = NULL,
  start_date = NULL,
  end_date = NULL,
  n_col,
  name_output,
  res = 300,
  height = 8,
  width = 6,
  ticks_x = 0.1,
  ticks_y = 0.1,
  n_cores = 1,
  xlim,
  ylim,
  save_data = TRUE,
  from_data = FALSE,
  data_plot_file = NULL,
  sensor
)
}
\arguments{
\item{dir_input}{directorio en donde se almacenan las tablas en csv o parquet}

\item{season}{temporalidad para la generación de imágenes en formato raster("mes", o "year")}

\item{stat_function}{función estadística para generar las imágenes raster ("median" o "mean")}

\item{var_name}{nombre de la variable a analizar ("chlor_a", "sst", "Rrs_645", "pic", "poc", "nflh", etc)}

\item{shp_file}{nombre archivo shp para el gráfico (si no esta en dir_input poner nombre con ruta completa). Opcional}

\item{start_date}{fecha inicio (formato "YYYYY-MM-DD")}

\item{end_date}{fecha final}

\item{n_col}{numero de columnas para el gráfico}

\item{name_output}{nombre para la salidas}

\item{res}{resolución para la imágen png de la climatología}

\item{height}{altura para la imágen png de la climatología}

\item{width}{amplitud para la imágen png de la climatología}

\item{ticks_x}{vector numérico tamaño 1 indicando ticks en el eje x del plot}

\item{ticks_y}{vector numérico tamaño 1 indicando ticks en el eje y del plot}

\item{n_cores}{cores a usar}

\item{xlim}{vector numérico tamaño 2 para los limites de x en el plot}

\item{ylim}{vector numérico tamaño 2 para los limites de y en el plot}

\item{save_data}{opcion boleana}

\item{from_data}{opcion boleana}

\item{data_plot_file}{nombre de archivo}

\item{sensor}{nombre del sensor (solo sirve para la leyenda)}
}
\value{
imágenes png
}
\description{
Función para generar climatología y exportarla (formato png) para una variable determinada
}
\examples{
\dontrun{
dir_input <- "/home/holon--oim/Dropbox/HolonSPA/PROYECTOS/Proyecto_Junin/Analisis/satelital/aqua/chlor_a/parquet_files"
season <- "month"
stat_function <- "median"
var_name <- "chlor_a"
#shp_file <- "/home/holon--oim/Dropbox/HolonSPA/PROYECTOS/Proyecto_Junin/Analisis/satelital/zona_proyecto_junin.shp"
n_col <- 4
name_output <- "chlor_aqua_proyecto_junin.png"
ticks_x = 0.1
ticks_y = 0.15
n_cores = 16
xlim <- c(west, east)
ylim <- c(north, south)
sensor = "aqua"
plot_clim(dir_input = dir_input,season = season, stat_function = stat_function,var_name = var_name,n_col = n_col,name_output = name_output,res = 300,height = 12,width = 9,ticks_x = ticks_x, ticks_y = ticks_y,n_cores = n_cores,xlim = xlim, ylim = ylim,save_data = FALSE,from_data = TRUE,data_plot_file = "chlor_aqua_proyecto_junin.csv", sensor = sensor)
}
}
