% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-clim.R
\name{plot_clim}
\alias{plot_clim}
\title{plot_clim}
\usage{
plot_clim(
  dir_input,
  season,
  stat_function,
  var_name,
  shp_file,
  n_col,
  name_output,
  res = 300,
  height = 8,
  width = 6,
  ticks_x = 0.2,
  ticks_y = 0.1,
  n_cores = 1
)
}
\arguments{
\item{dir_input}{directorio en donde se almacenan las tablas en csv o parquet}

\item{season}{temporalidad para la generación de imágenes en formato raster("mes", o "year")}

\item{stat_function}{función estadística para generar las imágenes raster ("median" o "mean")}

\item{var_name}{nombre de la variable a analizar ("chlor_a", "sst", "Rrs_645", "pic", "poc", "nflh", etc)}

\item{shp_file}{nombre archivo shp para el gráfico (si no esta en dir_input poner nombre con ruta completa)}

\item{n_col}{numero de columnas para el gráfico}

\item{name_output}{nombre para la salidas}

\item{res}{resolución para la imágen png de la climatología}

\item{height}{altura para la imágen png de la climatología}

\item{width}{amplitud para la imágen png de la climatología}

\item{ticks_x}{vector numérico tamaño 1 indicando ticks en el eje x del plot}

\item{ticks_y}{vector numérico tamaño 1 indicando ticks en el eje y del plot}

\item{xlim}{vector numérico tamaño 2 para los limites de x en el plot}

\item{ylim}{vector numérico tamaño 2 para los limites de y en el plot}
}
\value{
imágenes png
}
\description{
Función para generar climatología y exportarla (formato png) para una variable determinada
}
\examples{
\dontrun{
dir_input <- "/dir/to/input"
season <- "month"
stat_function <- "median"
var_name <- "chlor_a"
n_col <- 3
name_output <- "climatologia_chlor_a.png"
shp_file <- "/home/evolecolab/Escritorio/test_satImg/raster_mensuales/resultados_raster/Golfo_Arauco_prj2.shp"
res <- 300
height <- 7
width <- 9
plot_clim(dir_input = dir_input, season = season, stat_function = stat_function, var_name = var_name, shp_file = shp_file, n_col = n_col, n_cores, name_output = name_output, res = res, height = height, width = width)
}
}
