% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-facet-oce.R
\name{plot_facet_oce}
\alias{plot_facet_oce}
\title{plot_facet_oce}
\usage{
plot_facet_oce(
  list_nc,
  var,
  name_plot,
  height,
  width,
  shp_file,
  start_time,
  end_time
)
}
\arguments{
\item{var}{string que indica el nombre de la variable a graficar}

\item{name_plot}{string que indica el nombre de la imagen de salida}

\item{height}{altura del plot}

\item{width}{ancho del plot}

\item{shp_file}{nombre del shp utilizado para la imagen final}

\item{start_time}{fecha de inicio del el plot}

\item{end_time}{fecha de termino del plot}

\item{lists_nc}{lista de archivos nc de las variables}
}
\value{
imágenes png parámetros oceanográficos
}
\description{
Función para obtener imágenes del informe mensual para cada parametro oceanográfico (chlor_a, sst y rrs_645)
}
\examples{
\dontrun{
shp_file <- "/media/evolecolab/PortableSSD/seguimiento_arauco_2022/oce_pars/Golfo_Arauco_prj2.shp"
height <- 5
width <- 14
lists_nc <- dir_ls(path = "/media/evolecolab/PortableSSD/seguimiento_arauco_2022/oce_pars/var/input_plot/semana_20", regexp = ".nc$")
var <- "sst"
start_time <- "2023-03-20"
end_time <- "2023-03-26"
name_plot <- "ocepars_plot_w20.png"
plot_facet_oce(lists_nc = lists_nc, var = var, name_plot = name_plot, height = height, width = width, shp_file = shp_file, start_time = start_time, end_time = end_time)
}
}
